+++ src/main/java/com/snack/news/domain/topic/Topic.java
@@ -1,4 +1,4 @@
-package com.snack.news.domain;
+package com.snack.news.domain.topic;
 
 import lombok.Builder;
 import lombok.Getter;
