+++ src/main/java/com/snack/news/domain/topic/TopicType.java
@@ -1,4 +1,4 @@
-package com.snack.news.domain;
+package com.snack.news.domain.topic;
 
 public enum TopicType {
 	CORP, PERSON, FIELD, NONE
