+++ src/main/java/com/snack/news/dto/TopicDto.java
@@ -1,7 +1,7 @@
 package com.snack.news.dto;
 
-import com.snack.news.domain.Topic;
-import com.snack.news.domain.TopicType;
+import com.snack.news.domain.topic.Topic;
+import com.snack.news.domain.topic.TopicType;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
