+++ src/test/java/com/snack/news/repository/NewsRepositoryTest.java
@@ -3,7 +3,7 @@ package com.snack.news.repository;
 import com.snack.news.domain.Category;
 import com.snack.news.domain.News;
 import com.snack.news.domain.Tag;
-import com.snack.news.domain.Topic;
+import com.snack.news.domain.topic.Topic;
 import com.snack.news.dto.NewsDto;
 import com.snack.news.fixture.NewsFixture;
 import org.junit.Test;
